requirements.txt
openai
tiktoken 
chromadb
langchain
flask
GitPython
python-dotenv
pydantic>=2.0